test(omo-senpi): split the memorian wiring suite by behaviour
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
